refactor: group interlinks inventory and index operations
Move the Sphinx inventory format into the interlinks package that
consumes it. Keep index serialisation beside project-level index
construction. Explain how documented objects claim short names and how
unique and ambiguous claims are handled.
